Moved archive into it's own package, updated references and renamed functions and internal structs to fit the move. Updated a few log locations to use correct package and in some cases return error instead of panic.
